♻ Moved bounding box logic
As Bounding-Box is related to an event list, it maes more sense to have
its logic in `event.rs` than in a specific compilation pass.
